feat(properties): log properties information
