+++ src/properties/map_test.go
